+++ report-front-end/package.json
@@ -12,50 +12,37 @@
     "@cloudscape-design/components": "^3.0.648",
     "@cloudscape-design/design-tokens": "^3.0.28",
     "@cloudscape-design/global-styles": "^1.0.13",
-    "@fortawesome/fontawesome-svg-core": "^6.4.2",
-    "@fortawesome/free-solid-svg-icons": "^6.4.2",
-    "@fortawesome/react-fontawesome": "^0.2.0",
     "aws-amplify": "^5.3.12",
     "dotenv": "^16.4.5",
-    "luxon": "^3.4.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hot-toast": "^2.4.1",
-    "react-json-view-lite": "^0.9.8",
     "react-markdown": "^9.0.0",
     "react-redux": "^9.1.2",
     "react-router-dom": "^6.15.0",
-    "react-speech-recognition": "^3.10.0",
     "react-syntax-highlighter": "^15.5.0",
     "react-textarea-autosize": "^8.5.3",
     "react-use-websocket": "^4.8.1",
     "redux": "^5.0.1",
     "regenerator-runtime": "^0.14.0",
-    "remark-gfm": "^4.0.0",
     "umi-request": "^1.4.0",
     "uuid": "^9.0.0"
   },
   "devDependencies": {
-    "@types/luxon": "^3.3.2",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
-    "@types/react-speech-recognition": "^3.9.2",
     "@types/react-syntax-highlighter": "^15.5.13",
     "@types/uuid": "^9.0.3",
-    "@types/zen-observable": "^0.8.5",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
     "@vitejs/plugin-react": "^4.0.3",
     "autoprefixer": "^10.4.14",
-    "classnames": "^2.5.1",
-    "cross-env": "^7.0.3",
     "eslint": "^8.45.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.3",
     "postcss": "^8.4.27",
     "sass": "^1.65.1",
     "typescript": "^5.0.2",
-    "vite": "^4.5.2",
-    "zen-observable": "^0.10.0"
+    "vite": "^4.5.2"
   }
 }
