+++ report-front-end/src/app.tsx
@@ -1,17 +1,20 @@
 import { AmplifyUser } from "@aws-amplify/ui";
 import { UseAuthenticator } from "@aws-amplify/ui-react-core";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import toast, { Toaster } from "react-hot-toast";
 import { useDispatch } from "react-redux";
-import { BrowserRouter } from "react-router-dom";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+import { v4 as uuid } from "uuid";
 import "./app.scss";
-import {
-  isLoginWithCognito,
-  LOCAL_STORAGE_KEYS,
-} from "./common/constant/constants";
-import { ActionType, UserInfo } from "./common/helpers/types";
-import CustomTopNavigation from "./components/top-navigation";
-import PageRouter from "./pages/page-router";
+import BaseAppLayout from "./components/BaseAppLayout";
+import PanelConfigs from "./components/PanelConfigs";
+import { PanelSideNav } from "./components/PanelSideNav";
+import { Session } from "./components/PanelSideNav/types";
+import SectionChat from "./components/SectionChat";
+import TopNav from "./components/TopNav";
+import { GlobalContext } from "./hooks/useGlobalContext";
+import { isLoginWithCognito, LOCAL_STORAGE_KEYS } from "./utils/constants";
+import { ActionType, UserInfo } from "./utils/helpers/types";
 
 export type SignOut = UseAuthenticator["signOut"];
 
@@ -67,10 +70,12 @@ const App: React.FC<{
     <div style={{ height: "100%" }}>
       <BrowserRouter>
         <Toaster />
-        <CustomTopNavigation />
+        <TopNav />
         <div style={{ height: "56px", backgroundColor: "#000716" }}>&nbsp;</div>
         <div>
-          <PageRouter />
+          <Routes>
+            <Route index path="/" element={<Playground />} />
+          </Routes>
         </div>
       </BrowserRouter>
     </div>
@@ -78,3 +83,30 @@ const App: React.FC<{
 };
 
 export default App;
+
+const initSession = () => ({
+  session_id: uuid(),
+  title: "New Chat",
+  messages: [],
+});
+
+function Playground() {
+  const [toolsHide, setToolsHide] = useState(true);
+  const [sessions, setSessions] = useState<Session[]>([initSession()]);
+  const [currentSessionId, setCurrentSessionId] = useState(
+    sessions[0].session_id
+  );
+  return (
+    <GlobalContext.Provider
+      value={{ sessions, setSessions, currentSessionId, setCurrentSessionId }}
+    >
+      <BaseAppLayout
+        navigation={<PanelSideNav />}
+        content={<SectionChat {...{ toolsHide, setToolsHide }} />}
+        tools={<PanelConfigs setToolsHide={setToolsHide} />}
+        toolsHide={toolsHide}
+        setToolsHide={setToolsHide}
+      />
+    </GlobalContext.Provider>
+  );
+}
