--- report-front-end/src/common/types.ts
@@ -1,4 +1,0 @@
-export interface NavigationPanelState {
-  collapsed?: boolean;
-  collapsedSections?: Record<number, boolean>;
-}
