+++ report-front-end/src/components/Login/index.tsx
@@ -1,0 +1,6 @@
+import CognitoLogin from "./CognitoLogin";
+import CustomLogin from "./CustomLogin";
+
+const Login = { Cognito: CognitoLogin, Custom: CustomLogin };
+
+export default Login;
