+++ report-front-end/src/components/PanelConfigs/style.scss
