+++ report-front-end/src/components/PanelSideNav/style.scss
