+++ report-front-end/src/components/PanelSideNav/types.ts
@@ -1,0 +1,7 @@
+import { ChatBotHistoryItem } from "../SectionChat/types";
+
+export interface Session {
+  session_id: string;
+  title: string;
+  messages: ChatBotHistoryItem[];
+}
