+++ report-front-end/src/components/SectionChat/ExpandableSectionWithDivider.tsx
