+++ report-front-end/src/components/SectionChat/chat.module.scss
