--- report-front-end/src/components/session-panel/types.ts
@@ -1,7 +1,0 @@
-import { ChatBotHistoryItem } from "../chatbot-panel/types";
-
-export interface Session {
-  session_id: string;
-  title: string;
-  messages: ChatBotHistoryItem[];
-}
