+++ report-front-end/src/utils/constants.ts
