+++ report-front-end/src/utils/helpers/types.ts
